+++ DependencyInjection/HWIOAuthExtension.php
@@ -36,24 +36,29 @@ class HWIOAuthExtension extends Extension
         $loader->load('oauth.xml');
         $loader->load('templating.xml');
         $loader->load('twig.xml');
-        $loader->load('buzz.xml');
+        $loader->load('http_client.xml');
 
         $processor = new Processor();
         $config = $processor->processConfiguration(new Configuration(), $configs);
 
-        // setup buzz client settings
-        $httpClient = $container->getDefinition('buzz.client');
+        // setup http client settings
+        $httpClient = $container->getDefinition('hwi_oauth.http_client');
         $httpClient->addMethodCall('setVerifyPeer', array($config['http_client']['verify_peer']));
         $httpClient->addMethodCall('setTimeout', array($config['http_client']['timeout']));
         $httpClient->addMethodCall('setMaxRedirects', array($config['http_client']['max_redirects']));
         $httpClient->addMethodCall('setIgnoreErrors', array($config['http_client']['ignore_errors']));
-        $container->setDefinition('hwi_oauth.http_client', $httpClient);
+        if (isset($config['http_client']['proxy']) && $config['http_client']['proxy'] != '') {
+            $httpClient->addMethodCall('setProxy', array($config['http_client']['proxy']));
+        }
 
         // set current firewall
         $container->setParameter('hwi_oauth.firewall_name', $config['firewall_name']);
 
         // set target path parameter
         $container->setParameter('hwi_oauth.target_path_parameter', $config['target_path_parameter']);
+
+        // set use referer parameter
+        $container->setParameter('hwi_oauth.use_referer', $config['use_referer']);
 
         // setup services for all configured resource owners
         $resourceOwners = array();
@@ -65,6 +70,13 @@ class HWIOAuthExtension extends Extension
 
         $oauthUtils = $container->getDefinition('hwi_oauth.security.oauth_utils');
         $oauthUtils->addMethodCall('setResourceOwnerMap', array(new Reference('hwi_oauth.resource_ownermap.'.$config['firewall_name'])));
+        // Symfony <2.6 BC
+        // Go back to basic xml config after
+        if (interface_exists('Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface')) {
+            $oauthUtils->replaceArgument(1, new Reference('security.authorization_checker'));
+        } else {
+            $oauthUtils->replaceArgument(1, new Reference('security.context'));
+        }
 
         if (isset($config['fosub'])) {
             $container
@@ -134,7 +146,8 @@ class HWIOAuthExtension extends Extension
             $type = $options['type'];
             unset($options['type']);
 
-            $definition = new DefinitionDecorator('hwi_oauth.abstract_resource_owner.'.$type);
+            $definition = new DefinitionDecorator('hwi_oauth.abstract_resource_owner.'.Configuration::getResourceOwnerType($type));
+            $definition->setClass("%hwi_oauth.resource_owner.$type.class%");
             $container->setDefinition('hwi_oauth.resource_owner.'.$name, $definition);
             $definition
                 ->replaceArgument(2, $options)