+++ OAuth/ResourceOwner/EventbriteResourceOwner.php
@@ -12,7 +12,6 @@
 namespace HWI\Bundle\OAuthBundle\OAuth\ResourceOwner;
 
 use Buzz\Message\RequestInterface as HttpRequestInterface;
-use HWI\Bundle\OAuthBundle\Security\Core\Authentication\Token\OAuthToken;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 /**
@@ -28,28 +27,11 @@ class EventbriteResourceOwner extends GenericOAuth2ResourceOwner
     protected $paths = array(
         'identifier' => 'user.user_id',
         'nickname'   => 'user.first_name',
+        'firstname'  => 'user.first_name',
+        'lastname'   => 'user.last_name',
         'realname'   => array('user.first_name', 'user.last_name'),
         'email'      => 'email',
     );
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getUserInformation(array $accessToken, array $extraParameters = array())
-    {
-        $url = $this->normalizeUrl($this->getOption('infos_url'), array(
-            'access_token' => $accessToken['access_token']
-        ));
-
-        $content = $this->httpRequest($url, null, array('Authorization: Bearer '.$accessToken['access_token']))->getContent();
-
-        $response = $this->getUserResponse();
-        $response->setResponse($content);
-        $response->setResourceOwner($this);
-        $response->setOAuthToken(new OAuthToken($accessToken));
-
-        return $response;
-    }
 
     /**
      * {@inheritDoc}
@@ -67,9 +49,11 @@ class EventbriteResourceOwner extends GenericOAuth2ResourceOwner
         parent::configureOptions($resolver);
 
         $resolver->setDefaults(array(
-            'authorization_url' => 'https://www.eventbrite.com/oauth/authorize',
-            'access_token_url'  => 'https://www.eventbrite.com/oauth/token',
-            'infos_url'         => 'https://www.eventbrite.com/json/user_get',
+            'authorization_url'        => 'https://www.eventbrite.com/oauth/authorize',
+            'access_token_url'         => 'https://www.eventbrite.com/oauth/token',
+            'infos_url'                => 'https://www.eventbrite.com/json/user_get',
+
+            'use_bearer_authorization' => true,
         ));
     }
 }