+++ OAuth/ResourceOwner/GenericOAuth2ResourceOwner.php
@@ -31,17 +31,14 @@ class GenericOAuth2ResourceOwner extends AbstractResourceOwner
     public function getUserInformation(array $accessToken, array $extraParameters = array())
     {
         if ($this->options['use_bearer_authorization']) {
-            $url = $this->normalizeUrl($this->options['infos_url']);
-
-            $content = $this->httpRequest($url, null, array('Authorization: Bearer '.$accessToken['access_token']));
+            $content = $this->httpRequest($this->normalizeUrl($this->options['infos_url']), null, array('Authorization: Bearer '.$accessToken['access_token']));
         } else {
-            $url = $this->normalizeUrl($this->options['infos_url'], array('access_token' => $accessToken['access_token']));
-
-            $content = $this->doGetUserInformationRequest($url);
+            $content = $this->doGetUserInformationRequest($this->normalizeUrl($this->options['infos_url'], array($this->options['attr_name'] => $accessToken['access_token'])));
         }
 
         $response = $this->getUserResponse();
         $response->setResponse($content->getContent());
+
         $response->setResourceOwner($this);
         $response->setOAuthToken(new OAuthToken($accessToken));
 
@@ -180,7 +177,7 @@ class GenericOAuth2ResourceOwner extends AbstractResourceOwner
      */
     protected function doGetUserInformationRequest($url, array $parameters = array())
     {
-        return $this->httpRequest($url);
+        return $this->httpRequest($url, http_build_query($parameters, '', '&'));
     }
 
     /**
@@ -211,28 +208,41 @@ class GenericOAuth2ResourceOwner extends AbstractResourceOwner
         parent::configureOptions($resolver);
 
         $resolver->setDefaults(array(
+            'attr_name'                => 'access_token',
             'use_commas_in_scope'      => false,
             'use_bearer_authorization' => true,
         ));
 
-        $resolver->setOptional(array(
-            'revoke_token_url',
-        ));
+        // Symfony <2.6 BC
+        if (method_exists($resolver, 'setDefined')) {
+            $resolver->setDefined('revoke_token_url');
+        } else {
+            $resolver->setOptional(array(
+                'revoke_token_url',
+            ));
+        }
 
-        $resolver->setNormalizers(array(
-            // Unfortunately some resource owners break the spec by using commas instead
-            // of spaces to separate scopes (Disqus, Facebook, Github, Vkontante)
-            'scope' => function (Options $options, $value) {
-                if (!$value) {
-                    return null;
-                }
+        // Unfortunately some resource owners break the spec by using commas instead
+        // of spaces to separate scopes (Disqus, Facebook, Github, Vkontante)
+        $scopeNormalizer = function (Options $options, $value) {
+            if (!$value) {
+                return null;
+            }
 
-                if (!$options['use_commas_in_scope']) {
-                    return $value;
-                }
+            if (!$options['use_commas_in_scope']) {
+                return $value;
+            }
 
-                return str_replace(',', ' ', $value);
-            },
-        ));
+            return str_replace(',', ' ', $value);
+        };
+
+        // Symfony <2.6 BC
+        if (method_exists($resolver, 'setNormalizer')) {
+            $resolver->setNormalizer('scope', $scopeNormalizer);
+        } else {
+            $resolver->setNormalizers(array(
+                'scope' => $scopeNormalizer,
+            ));
+        }
     }
 }