+++ OAuth/Response/PathUserResponse.php
@@ -26,6 +26,8 @@ class PathUserResponse extends AbstractUserResponse
     protected $paths = array(
         'identifier'     => null,
         'nickname'       => null,
+        'firstname'      => null,
+        'lastname'       => null,
         'realname'       => null,
         'email'          => null,
         'profilepicture' => null,
@@ -45,6 +47,22 @@ class PathUserResponse extends AbstractUserResponse
     public function getNickname()
     {
         return $this->getValueForPath('nickname');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFirstName()
+    {
+        return $this->getValueForPath('firstname');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLastName()
+    {
+        return $this->getValueForPath('lastname');
     }
 
     /**
@@ -96,7 +114,7 @@ class PathUserResponse extends AbstractUserResponse
      *
      * @return mixed
      */
-    protected function getPath($name)
+    public function getPath($name)
     {
         return array_key_exists($name, $this->paths) ? $this->paths[$name] : null;
     }