+++ Tests/OAuth/ResourceOwner/GenericOAuth2ResourceOwnerTest.php
@@ -11,10 +11,13 @@
 
 namespace HWI\Bundle\OAuthBundle\Tests\OAuth\ResourceOwner;
 
+use Buzz\Exception\RequestException;
 use Buzz\Message\MessageInterface;
 use Buzz\Message\RequestInterface;
+use HWI\Bundle\OAuthBundle\OAuth\Exception\HttpTransportException;
 use HWI\Bundle\OAuthBundle\OAuth\ResourceOwner\GenericOAuth2ResourceOwner;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
 
 class GenericOAuth2ResourceOwnerTest extends \PHPUnit_Framework_TestCase
 {
@@ -38,6 +41,8 @@ class GenericOAuth2ResourceOwnerTest extends \PHPUnit_Framework_TestCase
         'infos_url'           => 'http://user.info/?test=1',
         'authorization_url'   => 'http://user.auth/?test=2',
         'access_token_url'    => 'http://user.access/?test=3',
+
+        'attr_name'           => 'access_token',
     );
 
     protected $userResponse = <<<json
@@ -65,15 +70,15 @@ json;
     }
 
     /**
-     * @expectedException \Symfony\Component\OptionsResolver\Exception\InvalidArgumentException
+     * @expectedException \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      */
-    public function testInvalidOptionThrowsException()
+    public function testUndefinedOptionThrowsException()
     {
         $this->createResourceOwner($this->resourceOwnerName, array('non_existing' => null));
     }
 
     /**
-     * @expectedException \Symfony\Component\OptionsResolver\Exception\InvalidArgumentException
+     * @expectedException \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      */
     public function testInvalidOptionValueThrowsException()
     {
@@ -94,6 +99,22 @@ json;
         $this->assertEquals('token', $userResponse->getAccessToken());
         $this->assertNull($userResponse->getRefreshToken());
         $this->assertNull($userResponse->getExpiresIn());
+    }
+
+    public function testGetUserInformationFailure()
+    {
+        $exception = new RequestException();
+
+        $this->buzzClient->expects($this->once())
+            ->method('send')
+            ->will($this->throwException($exception));
+
+        try {
+            $this->resourceOwner->getUserInformation(array('access_token' => 'token'));
+            $this->fail('An exception should have been raised');
+        } catch (HttpTransportException $e) {
+            $this->assertSame($exception, $e->getPrevious());
+        }
     }
 
     public function testGetAuthorizationUrl()
@@ -299,7 +320,7 @@ json;
         $this->storage->expects($this->once())
             ->method('fetch')
             ->with($resourceOwner, 'invalid_token', 'csrf_state')
-            ->will($this->throwException(new \InvalidArgumentException('No data available in storage.')));
+            ->will($this->throwException(new InvalidOptionsException('No data available in storage.')));
 
         $resourceOwner->isCsrfTokenValid('invalid_token');
     }